+++ src/server/httpServer.js
@@ -1,5 +1,4 @@
 import path                 from 'path';
-import fs                   from 'fs';
 import http                 from 'http';
 import Promise              from 'bluebird';
 import { cloneDeep }        from 'lodash';
@@ -11,6 +10,8 @@ import graphqlHttp          from 'express-graphql';
 import ejs                  from 'ejs';
 import cookieParser         from 'cookie-parser';
 import compression          from 'compression';
+import addAllLocales        from '../locales/all';
+import _t                   from '../translate';
 import * as gqlServer       from './gqlServer';
 let webpack;
 let webpackDevMiddleware;
@@ -39,7 +40,6 @@ try {
 /* eslint-enable global-require */
 
 const ASSET_PATH = '../../public';
-const LOCALE_PATH = '../locales';
 const DEFAULT_BOOTSTRAP = {
   ssrHtml: '',
   ssrCss: '',
@@ -49,25 +49,23 @@ const DEFAULT_BOOTSTRAP = {
 };
 const COOKIE_NAMESPACE = 'mady';
 
+addAllLocales();
+
 function sendIndexHtml(req, res) {
   mainStory.info('http', 'Preparing index.html...');
-  let userLang = req.query.lang || req.cookies[`${COOKIE_NAMESPACE}_lang`] || 'en';
+  const userLang = req.query.lang || req.cookies[`${COOKIE_NAMESPACE}_lang`] || 'en';
   const bootstrap = cloneDeep(DEFAULT_BOOTSTRAP);
   return Promise.resolve()
 
     // Locales
     .then(() => {
-      let langPath = path.join(__dirname, LOCALE_PATH, `${userLang}.js`);
-      mainStory.debug('http', `Reading ${chalk.cyan.bold(langPath)}...`);
-      try {
-        bootstrap.fnLocales = fs.readFileSync(langPath, 'utf8');
-        bootstrap.jsonData.lang = userLang;
-      } catch (err) {
-        userLang = 'en';
-        langPath = path.join(__dirname, LOCALE_PATH, `${userLang}.js`);
-        mainStory.debug('http', `Not found. Reading ${chalk.cyan.bold(langPath)} instead...`);
-        bootstrap.fnLocales = fs.readFileSync(langPath, 'utf8');
-        bootstrap.jsonData.lang = userLang;
+      mainStory.debug('http', `Getting locale code for lang ${chalk.cyan.bold(userLang)}...`);
+      const { lang, result } = _t.getLocaleCode(userLang);
+      bootstrap.fnLocales = result;
+      bootstrap.jsonData.lang = lang;
+      if (lang && lang !== userLang) {
+        mainStory.info('http',
+          `Serving locales for ${chalk.cyan.bold(lang)} instead of ${chalk.cyan.bold(userLang)}`);
       }
     })
 
@@ -76,7 +74,7 @@ function sendIndexHtml(req, res) {
       if (!ssr) return null;
       mainStory.debug('http', 'Rendering...');
       return ssr.render(req, {
-        lang: userLang,
+        lang: bootstrap.jsonData.lang,
         fnLocales: bootstrap.fnLocales,
       })
       .then(results => {